+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 Changelog
 
-## Unreleased
+## 0.2.0 - 2023-12-12
 
 - Remove identity support. Use Context instead. (#30)
 - Re-fetch when context attributes change. (#31)